+++ crates/node/src/main.rs
@@ -202,7 +202,7 @@ async fn main() -> anyhow::Result<()> {
 
     app.start_and_drive_to_end().await?;
 
-    tracing::info!("Shutting down analytics");
+    tracing::info!("🔌 Shutting down analytics...");
     let _ = analytics.shutdown();
 
     Ok(())